test: member 테스트 추가
